refactor: Modularize sign-in logic by creating SigninClient component and streamline provider handling
Only enabled / configured providers buttons will be shown.
